feat: rescope poi-radio
