+++ .github/workflows/docker.yml
@@ -9,7 +9,7 @@ on:
 
 env:
   REGISTRY: ghcr.io
-  BASE_NAME: graphops/poi-radio
+  BASE_NAME: graphops/subgraph-radio
 
 jobs:
   build-linux:
