+++ one-shot/Cargo.toml
@@ -5,13 +5,13 @@ edition = "2021"
 authors = ["GraphOps (axiomatic-aardvark, hopeyen)"]
 description = "One-shot messaging using Graphcast SDK (can be separated to a different repo)"
 license = "Apache-2.0"
-repository = "https://github.com/graphops/poi-radio"
+repository = "https://github.com/graphops/subgraph-radio"
 keywords = ["graphprotocol", "data-integrity", "Indexer", "waku", "p2p"]
 categories = ["network-programming", "web-programming::http-client"]
 
 [dependencies]
 graphcast-sdk = "0.4.0"
-poi-radio = { path = "../poi-radio" }
+subgraph-radio = { path = "../subgraph-radio" }
 prost = "0.11"
 once_cell = "1.17"
 chrono = "0.4"
