+++ subgraph-radio/src/graphql/schema_graph_node.graphql
