+++ test-runner/src/invalid_payload.rs
@@ -1,4 +1,4 @@
-use poi_radio::state::PersistedState;
+use subgraph_radio::state::PersistedState;
 use test_utils::{
     config::{test_config, TestSenderConfig},
     dummy_msg::DummyMsg,
@@ -43,9 +43,9 @@ pub async fn invalid_payload_test() {
 
     teardown(process_manager, &store_path);
 
-    let remote_messages = persisted_state.remote_messages();
+    let remote_ppoi_messages = persisted_state.remote_ppoi_messages();
     assert!(
-        remote_messages.is_empty(),
+        remote_ppoi_messages.is_empty(),
         "Remote messages should be empty"
     );
 }
